+++ plugin.php
@@ -4,7 +4,7 @@
  *
  * @wp-plugin
  *
- * Version: 170221.28901
+ * Version: 170221.48222
  * Text Domain: wp-sharks-core
  * Plugin Name: WP Sharks Core
  *